feat: 공감수 타입 변경

+++ src/main/java/issueissyu/backend/domain/pin/entity/Pin.java
@@ -30,7 +30,7 @@ public class Pin extends BaseEntity {
 
     @Builder.Default
     @Column(name = "like_count", nullable = false)
-    private long likeCount = 0L;
+    private int likeCount = 0;
 
     @Column(name = "pin_title", nullable = false, length = 100)
     private String pinTitle;
